backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.4
python-multipart==0.0.9
numpy==1.26.4
opencv-contrib-python-headless==4.10.0.84
Pillow==10.4.0
torch==2.3.1
mediapipe==0.10.14
pytest==8.2.2
httpx==0.27.0
